feat(api): Add onboardUrl API
Similar to `onboard` API, but returns URL instead of open the URL immediately.
