+++ src/api/onboard.ts
@@ -1,78 +1,13 @@
-import { stringify } from 'qs';
-
-import {
-  constructScopes,
-  generateConfigs,
-  mergeConfigs,
-  getIsTabValue,
-  generateCodeChallenge,
-  openWindow
-} from '../helper';
-import { MY_ACCOUNT_DOMAINS } from '../server-paths';
+import { getIsTabValue, openWindow } from '../helper';
 import { StoredOptions, OnboardOptions } from '../typings';
-import storage from '../storage';
+import onboardUrl from './onboard-url';
 
 export default function onboard(storedOptions: StoredOptions, options: OnboardOptions = {}): void {
   if (!window) {
     throw new Error('[mt-link-sdk] `onboard` only works in the browser.');
   }
 
-  const {
-    mode,
-    clientId,
-    cobrandClientId,
-    locale,
-    scopes: defaultScopes,
-    redirectUri: defaultRedirectUri
-  } = storedOptions;
+  const { isNewTab, ...restOptions } = options;
 
-  if (!clientId) {
-    throw new Error('[mt-link-sdk] Make sure to call `init` before calling `onboard`.');
-  }
-
-  const {
-    scopes = defaultScopes,
-    redirectUri = defaultRedirectUri,
-    pkce = false,
-    codeChallenge,
-    isNewTab,
-    state,
-    ...rest
-  } = options;
-
-  const configs = mergeConfigs(storedOptions, rest, ['authAction', 'showAuthToggle', 'showRememberMe', 'forceLogout']);
-
-  if (!redirectUri) {
-    throw new Error(
-      '[mt-link-sdk] Missing option `redirectUri` in `onboard`, make sure to pass one via `onboard` options or `init` options.'
-    );
-  }
-
-  const { email } = configs;
-
-  if (!email) {
-    throw new Error(
-      '[mt-link-sdk] Missing option `email` in `onboard`, make sure to pass one via `onboard` options or `init` options.'
-    );
-  }
-
-  storage.del('cv');
-
-  const cc = codeChallenge || (pkce && generateCodeChallenge());
-
-  const queryString = stringify({
-    client_id: clientId,
-    cobrand_client_id: cobrandClientId,
-    response_type: 'code',
-    scope: constructScopes(scopes),
-    redirect_uri: redirectUri,
-    code_challenge: cc || undefined,
-    code_challenge_method: cc ? 'S256' : undefined,
-    state,
-    country: 'JP',
-    locale,
-    configs: generateConfigs(configs)
-  });
-
-  openWindow(`${MY_ACCOUNT_DOMAINS[mode]}/onboard?${queryString}`, getIsTabValue(isNewTab));
+  openWindow(onboardUrl(storedOptions, restOptions), getIsTabValue(isNewTab));
 }
